Only explicitly run callback if 'URI malformed'
For some reason when a 'URI malformed' error is thrown the callback
never gets ran.  Initially I had this set to always run the callback
but noticed that with other errors it gets called correctly.

Obviously something else is going on here with this type of error causing
issues so there is probably a better fix available but for now this works.